DRY names CRUD forms
