+++ app/forms/steps/children/names_form.rb
@@ -1,36 +1,10 @@
 module Steps
   module Children
-    class NamesForm < BaseForm
-      attribute :names_attributes, Hash
-      attribute :new_name, StrippedString
-
-      validates_presence_of :new_name, if: :first_name?
-
+    class NamesForm < BaseNamesForm
       private
 
-      def persist!
-        raise C100ApplicationNotFound unless c100_application
-
-        create_new_child_with_name
-        update_existing_children_names
-
-        true
-      end
-
-      def create_new_child_with_name
-        record_collection.create(name: new_name) unless new_name.blank?
-      end
-
-      def update_existing_children_names
-        names_attributes.each_value do |attrs|
-          next if attrs.fetch('name').blank?
-          record_collection.update(attrs.fetch('id'), attrs)
-        end
-      end
-
-      def first_name?
-        return if c100_application.nil?
-        record_collection.empty?
+      def name_attribute
+        :name
       end
 
       def record_collection
